Moved packedbubble series into a subfolder.
